update index.html
Hey, so I just noticed a super tiny typo, I couldn't help myself. On lines 64-66: it was previously written, " <p>Ideally we're looking for a great female hacker who known what it's like to be a student in an American public
        school.</p>" 

I am pulling a request for "known" to be change to "knows". So sorry in advance if this is not allowed. I have no idea about this beforehand and this is totally not a major change or anything at all.. I just couldn't help myself when I see tiny typos / spelling errors. You guys are doing wonderful work though, am excited for the new wave for Hack Club!! 

Warmly, 
Snigdha